Move common Activity logic into a new class
